feat(output): use "display_id" instead of "id"
The id set in an extractor can be overwritten by an other extractor later in the chain.
The final id can be very long, resulting in a too long filename.
"display_id" is not overwritten (at least by the generic extractor).
It’s working as before with youtube and twitch.